feat: 타임캡슐 조회 시 마스킹 처리 추가

+++ api/src/main/java/com/feelarchive/api/common/response/LocationDetail.java
@@ -1,7 +1,9 @@
 package com.feelarchive.api.common.response;
 
+import java.math.BigDecimal;
+
 public record LocationDetail(
     String address,
-    Double latitude,
-    Double longitude
+    BigDecimal latitude,
+    BigDecimal longitude
 ) {}
